requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pika==1.3.2
pydantic==2.10.0
grpcio==1.68.0
grpcio-tools==1.68.0
protobuf==5.29.0